% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{create_ggkegg}
\alias{create_ggkegg}
\title{Create ggkegg Object for KEGG Pathway Visualization}
\usage{
create_ggkegg(kegg_results, pathway_number, organism_code = "sce")
}
\arguments{
\item{kegg_results}{An enrichKEGG output from clusterProfiler (provided by user).}

\item{pathway_number}{The number representing the pathway in KEGG.}

\item{organism_code}{A string representing the organism code in KEGG (e.g., "sce").}
}
\value{
A ggkegg object.
}
\description{
Creates a ggkegg object from KEGG results using specified parameters.
}
