% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{createParalogPlots}
\alias{createParalogPlots}
\title{Create KEGG Pathway Graphs}
\usage{
createParalogPlots(graph_data, pathway_id, fc_column, juke = 20, title)
}
\arguments{
\item{graph_data}{A dataframe with KEGG pathway data processed for visualization.}

\item{pathway_id}{A string indicating the pathway ID for overlay.}

\item{fc_column}{A string naming the first fold change column.}

\item{juke}{A number for amount of space to separate names. Default is 20.}

\item{title}{A title for the plot.}
}
\value{
A ggplot object displaying the pathway graphs side by side.
}
\description{
Creates comparative graphs for gene expression data on KEGG pathways.
}
