% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_enrich_results.R
\docType{data}
\name{example_enrich_results}
\alias{example_enrich_results}
\title{An example output of enrichKEGG is provided in the `example_enrich_results` dataset.
The results are the top enrichments for the genes in the lrt_yeast example.
This data is used in vignette examples.}
\format{
An object of class \code{enrichResult} with 37 rows and 11 columns.
}
\usage{
example_enrich_results
}
\description{
An example output of enrichKEGG is provided in the `example_enrich_results` dataset.
The results are the top enrichments for the genes in the lrt_yeast example.
This data is used in vignette examples.
}
\keyword{datasets}
