% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{fetchKeggReactions}
\alias{fetchKeggReactions}
\title{Fetch KEGG Reactions}
\usage{
fetchKeggReactions()
}
\value{
A dataframe containing KEGG reactions with their descriptions.
}
\description{
Retrieves KEGG reaction mappings and formats them for use.
}
