% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{getPathwayNumber}
\alias{getPathwayNumber}
\title{Get Pathway Number to match row of enrichKEGG to KEGG pathway of interest}
\usage{
getPathwayNumber(kegg_results, pathway_id)
}
\arguments{
\item{kegg_results}{An enrichKEGG output from clusterProfiler (provided by user).}

\item{pathway_id}{A string specifying the pathway ID (provided by user).}
}
\value{
An integer representing the pathway number.
}
\description{
Retrieves the pathway number based on a pathway ID from KEGG data.
}
