% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{loadEnrichResults}
\alias{loadEnrichResults}
\title{load Input data (enrichKEGG object)}
\usage{
loadEnrichResults(enrich_results)
}
\arguments{
\item{enrich_results}{enrichResult Object from enrichKEGG.}
}
\value{
A dataframe with results for subsequent analysis.
}
\description{
Processes outputs from enrichKEGG.
}
