% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{loadGeneData}
\alias{loadGeneData}
\title{load Input data (enrichKEGG object)}
\usage{
loadGeneData(DE_results, FC_col_name)
}
\arguments{
\item{DE_results}{gene results object, either edgeR, DESeq2, limma, or data frame.}

\item{FC_col_name}{A string specifying the column name for log fold change.}
}
\value{
A dataframe with results for subsequent analysis.
}
\description{
Processes outputs from enrichKEGG.
}
