% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_yeast.R
\docType{data}
\name{lrt_yeast}
\alias{lrt_yeast}
\title{An example DGELRT object is provided in the `lrt_yeast` dataset.
This object contains the results of a differential gene expression analysis using the edgeR package.
logFC values correspond to the effect of a gene knockout on response to ethanol stress in yeast.}
\format{
An object of class \code{DGELRT} with 5615 rows and 4 columns.
}
\usage{
lrt_yeast
}
\description{
An example DGELRT object is provided in the `lrt_yeast` dataset.
This object contains the results of a differential gene expression analysis using the edgeR package.
logFC values correspond to the effect of a gene knockout on response to ethanol stress in yeast.
}
\keyword{datasets}
