% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paralogs-package.R
\docType{package}
\name{Paralogs-package}
\alias{Paralogs}
\alias{Paralogs-package}
\title{Paralogs: Visualizing Differential Expression of Paralogs in KEGG Gene Pathways}
\description{
This package provides functions to visualize the differential expression of paralogs in KEGG gene pathways. The package includes functions to download KEGG gene pathways, extract paralogs from the pathways, and visualize the differential expression of the paralogs. The package also includes functions to visualize the expression of paralogs in the context of the KEGG gene pathways.
}
\author{
\strong{Maintainer}: Carson Stacy \email{carsonstacy2@gmail.com}

}
\keyword{internal}
