% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{plotParalogs}
\alias{plotParalogs}
\title{Full Workflow for KEGG Pathway Visualization}
\usage{
plotParalogs(
  enrich_results,
  DE_results,
  organism_code = "sce",
  annotation_package = org.Sc.sgd.db,
  keytype = "ORF",
  pathway_code,
  fc_column = "logFC",
  juke = 20,
  title = NULL
)
}
\arguments{
\item{enrich_results}{The enrichKEGG results (from clusterProfiler).}

\item{DE_results}{The DE results (e.g. from edgeR).}

\item{organism_code}{The organism code for KEGG pathways (e.g., 'sce').}

\item{annotation_package}{The Bioconductor package for annotations relevant to the organism.}

\item{keytype}{The type of key used in the gene IDs (e.g., "SYMBOL", "ENTREZID", "ENSEMBL").}

\item{pathway_code}{The KEGG pathway code to visualize (e.g. 'sce00020')}

\item{fc_column}{The first fold change column name for visualization.}

\item{juke}{A number for amount of space to separate names. Default is 20.}

\item{title}{A title for the plot.}
}
\value{
A ggplot object displaying the pathway graphs.
}
\description{
Executes a complete workflow from raw DE results to visualizing them on a KEGG pathway.
}
