% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{loadGeneData}
\alias{loadGeneData}
\title{load Input data (enrichKEGG object)}
\usage{
loadGeneData(DE_results, logFC)
}
\arguments{
\item{enrich_results}{enrichResult Object from enrichKEGG.}
}
\value{
A dataframe with results for subsequent analysis.
}
\description{
Processes outputs from enrichKEGG.
}
