% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paralogs.R
\name{process_ggkegg}
\alias{process_ggkegg}
\title{Process ggkegg Data for Visualization}
\usage{
process_ggkegg(
  KEGG_data,
  topTags_results,
  KEGG_reactions_df,
  annotation_db,
  organism_code = "sce"
)
}
\arguments{
\item{KEGG_data}{A ggkegg object containing KEGG pathway data.}

\item{topTags_results}{A dataframe containing gene results.}

\item{KEGG_reactions_df}{A dataframe containing KEGG reaction descriptions.}

\item{annotation_db}{An AnnotationDbi object for gene annotations.}

\item{organism_code}{A string representing the organism code in KEGG (e.g., "sce").}
}
\value{
A dataframe ready for plotting.
}
\description{
Processes ggkegg object data to prepare it for visualization, enhancing gene data with KEGG reactions and annotations.
}
